Move IUpdateCaloriesClient to another folder
